# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Decimal": {
				"type": "string"
			},
...
			"_36_Enums.LabStatus": {
				"type": "string",
				"enum": [
					"COMPLETED",
					"CANCELLED",
					"ORDERED",
					"SAMPLE_COLLECTED",
					"IN_PROGRESS",
					"VERIFIED"
				]
			},
			"OrderLabServiceDTO": {
				"properties": {
					"providedServiceId": {
						"type": "string"
					},
					"visitId": {
						"type": "string"
					},
					"specimenType": {
						"type": "string"
					}
				},
				"required": [
					"providedServiceId",
					"visitId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LabStatus": {
				"$ref": "#/components/schemas/_36_Enums.LabStatus"
			},
			"RecordLabResultDTO": {
				"properties": {
					"resultValue": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					},
					"referenceRange": {
						"type": "string"
					},
					"findings": {
						"type": "string"
					},
					"specimenType": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/LabStatus"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"_36_Enums.InvoiceStatus": {
				"type": "string",
				"enum": [
					"CANCELLED",
					"UNPAID",
					"PARTIALLY_PAID",
					"PAID"
				]
			},
			"InvoiceStatus": {
				"$ref": "#/components/schemas/_36_Enums.InvoiceStatus"
			},
			"_36_Enums.ChargeType": {
				"type": "string",
				"enum": [
					"WARD_STAY",
					"CONSULTATION",
					"LAB_TEST",
					"PHARMACY",
					"PROCEDURE",
					"MISCELLANEOUS"
				]
			},
			"ChargeType": {
				"$ref": "#/components/schemas/_36_Enums.ChargeType"
			},
			"InvoiceItemResponseDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"invoiceId": {
						"type": "string"
					},
					"chargeType": {
						"$ref": "#/components/schemas/ChargeType"
					},
					"referenceId": {
						"type": "string",
						"nullable": true
					},
					"description": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"invoiceId",
					"chargeType",
					"referenceId",
					"description",
					"quantity",
					"unitPrice",
					"totalPrice",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.PaymentMethod": {
				"type": "string",
				"enum": [
					"CASH",
					"MOBILE_MONEY",
					"INSURANCE",
					"CARD"
				]
			},
			"PaymentMethod": {
				"$ref": "#/components/schemas/_36_Enums.PaymentMethod"
			},
			"_36_Enums.PaymentStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"CANCELLED",
					"PARTIALLY_PAID",
					"PAID",
					"REFUNDEDInvoiceStatus"
				]
			},
			"PaymentStatus": {
				"$ref": "#/components/schemas/_36_Enums.PaymentStatus"
			},
			"PaymentResponseDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"receiptNumber": {
						"type": "string"
					},
					"invoiceId": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"status": {
						"$ref": "#/components/schemas/PaymentStatus"
					},
					"transactionRef": {
						"type": "string",
						"nullable": true
					},
					"receivedById": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"receiptNumber",
					"invoiceId",
					"amount",
					"paymentMethod",
					"status",
					"transactionRef",
					"receivedById",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvoiceResponseDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"invoiceNumber": {
						"type": "string"
					},
					"visitId": {
						"type": "string"
					},
					"patientId": {
						"type": "string"
					},
					"subtotal": {
						"type": "number",
						"format": "double"
					},
					"tax": {
						"type": "number",
						"format": "double"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"insurancePay": {
						"type": "number",
						"format": "double"
					},
					"patientPay": {
						"type": "number",
						"format": "double"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"amountPaid": {
						"type": "number",
						"format": "double"
					},
					"balanceDue": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/InvoiceStatus"
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/InvoiceItemResponseDTO"
						},
						"type": "array"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/PaymentResponseDTO"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"invoiceNumber",
					"visitId",
					"patientId",
					"subtotal",
					"tax",
					"discount",
					"insurancePay",
					"patientPay",
					"totalAmount",
					"amountPaid",
					"balanceDue",
					"status",
					"notes",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecordPaymentDTO": {
				"properties": {
					"invoiceId": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"paymentMethod": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"transactionRef": {
						"type": "string",
						"nullable": true
					},
					"receivedById": {
						"type": "string"
					}
				},
				"required": [
					"invoiceId",
					"amount",
					"paymentMethod",
					"receivedById"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePaymentDTO": {
				"$ref": "#/components/schemas/RecordPaymentDTO"
			},
...
			"_36_Enums.WardType": {
				"type": "string",
				"enum": [
					"MALE",
					"FEMALE",
					"GENERAL",
					"PEDIATRIC",
					"ICU",
					"HDUK",
					"MATERNITY",
					"SURGICAL",
					"ISOLATION",
					"PRIVATE"
				]
			},
			"WardType": {
				"$ref": "#/components/schemas/_36_Enums.WardType"
			},
			"CreateWardDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/WardType"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"dailyRate": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"dailyRate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBedDTO": {
				"properties": {
					"wardId": {
						"type": "string"
					},
					"bedNumber": {
						"type": "string"
					}
				},
				"required": [
					"wardId",
					"bedNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdmitPatientDTO": {
				"properties": {
					"visitId": {
						"type": "string"
					},
					"patientId": {
						"type": "string"
					},
					"bedId": {
						"type": "string"
					},
					"admittedById": {
						"type": "string"
					},
					"admissionNotes": {
						"type": "string"
					}
				},
				"required": [
					"visitId",
					"patientId",
					"bedId",
					"admittedById"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DischargePatientDTO": {
				"properties": {
					"dischargedById": {
						"type": "string"
					},
					"dischargeNotes": {
						"type": "string"
					}
				},
				"required": [
					"dischargedById"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"description": "Authentication and RBAC Microservice",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/labs/order": {
			"post": {
				"operationId": "OrderLab",
				"responses": {
					"201": {
						"description": "Lab Service Ordered",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"properties": {
												"providedService": {
													"properties": {
														"unitPrice": {
															"$ref": "#/components/schemas/Decimal"
														},
														"providedById": {
															"type": "string"
														},
														"serviceId": {
															"type": "string"
														},
														"updatedAt": {
															"type": "string",
															"format": "date-time"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"notes": {
															"type": "string"
														},
														"visitId": {
															"type": "string"
														},
														"id": {
															"type": "string"
														}
													},
													"required": [
														"unitPrice",
														"providedById",
														"serviceId",
														"updatedAt",
														"createdAt",
														"notes",
														"visitId",
														"id"
													],
													"type": "object"
												}
											},
											"required": [
												"providedService"
											],
											"type": "object"
										},
										{
											"properties": {
												"verifiedById": {
													"type": "string"
												},
												"performedById": {
													"type": "string"
												},
												"findings": {
													"type": "string"
												},
												"referenceRange": {
													"type": "string"
												},
												"unit": {
													"type": "string"
												},
												"resultValue": {
													"type": "string"
												},
												"specimenType": {
													"type": "string"
												},
												"providedServiceId": {
													"type": "string"
												},
												"status": {
													"$ref": "#/components/schemas/_36_Enums.LabStatus"
												},
												"updatedAt": {
													"type": "string",
													"format": "date-time"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"visitId": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"verifiedById",
												"performedById",
												"findings",
												"referenceRange",
												"unit",
												"resultValue",
												"specimenType",
												"providedServiceId",
												"status",
												"updatedAt",
												"createdAt",
												"visitId",
												"id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Validation Error or Invalid Input"
					}
				},
				"description": "Order a new lab service for a patient visit.",
				"tags": [
					"Lab Results"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrderLabServiceDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/labs/{id}/results": {
			"put": {
				"operationId": "RecordResult",
				"responses": {
					"200": {
						"description": "Lab Result Recorded Successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"verifiedById": {
											"type": "string"
										},
										"performedById": {
											"type": "string"
										},
										"findings": {
											"type": "string"
										},
										"referenceRange": {
											"type": "string"
										},
										"unit": {
											"type": "string"
										},
										"resultValue": {
											"type": "string"
										},
										"specimenType": {
											"type": "string"
										},
										"providedServiceId": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.LabStatus"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"visitId": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"verifiedById",
										"performedById",
										"findings",
										"referenceRange",
										"unit",
										"resultValue",
										"specimenType",
										"providedServiceId",
										"status",
										"updatedAt",
										"createdAt",
										"visitId",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error or Invalid Status Transition"
					},
					"404": {
						"description": "Lab Result Not Found"
					}
				},
				"description": "Record findings, result values, and transition status to COMPLETED.",
				"tags": [
					"Lab Results"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecordLabResultDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/labs/{id}/verify": {
			"put": {
				"operationId": "VerifyResult",
				"responses": {
					"200": {
						"description": "Lab Result Verified Successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"verifiedById": {
											"type": "string"
										},
										"performedById": {
											"type": "string"
										},
										"findings": {
											"type": "string"
										},
										"referenceRange": {
											"type": "string"
										},
										"unit": {
											"type": "string"
										},
										"resultValue": {
											"type": "string"
										},
										"specimenType": {
											"type": "string"
										},
										"providedServiceId": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/_36_Enums.LabStatus"
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"visitId": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"verifiedById",
										"performedById",
										"findings",
										"referenceRange",
										"unit",
										"resultValue",
										"specimenType",
										"providedServiceId",
										"status",
										"updatedAt",
										"createdAt",
										"visitId",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Lab Result Not Found"
					}
				},
				"description": "Final verification of a lab result by a senior technician or doctor.\nTransitions status to VERIFIED.",
				"tags": [
					"Lab Results"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"findings": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/labs/visit/{visitId}": {
			"get": {
				"operationId": "GetByVisit",
				"responses": {
					"200": {
						"description": "Lab Results Retrieved",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve the complete lab history for a specific patient visit.",
				"tags": [
					"Lab Results"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "visitId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/invoices/{invoiceId}": {
			"get": {
				"operationId": "GetInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvoiceResponseDTO"
								}
							}
						}
					},
					"404": {
						"description": "Invoice not found"
					}
				},
				"description": "Get invoice details by ID, including line items and payment history",
				"tags": [
					"Billing & Invoices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "invoiceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/invoices/{invoiceId}/payments": {
			"post": {
				"operationId": "RecordPayment",
				"responses": {
					"201": {
						"description": "Payment Processed Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvoiceResponseDTO"
								}
							}
						}
					},
					"400": {
						"description": "Invalid payment amount or invoice already paid"
					},
					"404": {
						"description": "Invoice not found"
					}
				},
				"description": "Process a payment against an outstanding invoice",
				"tags": [
					"Billing & Invoices"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "invoiceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentDTO"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/admissions/wards": {
			"post": {
				"operationId": "CreateWard",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Register a new hospital Ward",
				"tags": [
					"Wards & Admissions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWardDTO"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllWards",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get all Wards (with linked Beds)",
				"tags": [
					"Wards & Admissions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/admissions/beds": {
			"post": {
				"operationId": "CreateBed",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Add a new Bed to a Ward",
				"tags": [
					"Wards & Admissions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBedDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/admissions/admit": {
			"post": {
				"operationId": "AdmitPatient",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Bed unavailable or bad request"
					}
				},
				"description": "Admit a patient to an available Bed and trigger RabbitMQ event",
				"tags": [
					"Wards & Admissions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdmitPatientDTO"
							}
						}
					}
				}
			}
		},
		"/api/v1/admissions/{admissionId}/discharge": {
			"put": {
				"operationId": "DischargePatient",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Active admission not found"
					}
				},
				"description": "Discharge an admitted patient, release Bed, and publish billing event",
				"tags": [
					"Wards & Admissions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "admissionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DischargePatientDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}